requirements.txt
pygame>=2.0
numpy
Pillow
requests
openai
diffusers
torch
